moved styled components to styled branch
