+++ src/components/galleryPage/galleryImage/GalleryImage.jsx
@@ -1,4 +1,4 @@
-import { StyledGalleryImage } from "./StyledGalleryImage";
+import { StyledGalleryImage } from "../../../styled/galleryPage/galleryImage/StyledGalleryImage";
 
 export const Image = ({ src, onOpenImageModal }) => (
   <StyledGalleryImage onClick={onOpenImageModal} src={src} />
