+++ src/components/galleryPage/modalInput/ModalInput.jsx
@@ -1,4 +1,4 @@
-import { StyledInput } from "./StyledModalInput";
+import { StyledInput } from "../../../styled/galleryPage/modalInput/StyledModalInput";
 
 export const ModalInput = ({ placeholder }) => (
   <StyledInput placeholder={placeholder} />
