+++ src/components/header/Header.jsx
@@ -1,3 +1,3 @@
-import { StyledHeader } from "./StyledHeader";
+import { StyledHeader } from "../../styled/header/StyledHeader";
 
 export const Header = () => <StyledHeader>Test app</StyledHeader>;
