+++ src/styled/footer/StyledFooter.jsx
