+++ src/styled/galleryPage/galleryImage/StyledGalleryImage.jsx
