+++ src/styled/galleryPage/modalButton/StyledModalButton.jsx
