+++ src/styled/galleryPage/modalInput/StyledModalInput.jsx
